[FIX] pos_stock,point_of_sale: restore rounding move line creation
Before this commit:
- The rounding move line creation was moved from point_of_sale to pos_stock
  while removing the dependency of stock on point_of_sale.
- As a result, when pos_stock was not installed, no rounding move lines
  were created for rounded PoS payments, leading to unbalanced journal
  entries during session closing.

In this commit:
- Restored the rounding move line creation in point_of_sale so that rounded
  payments are correctly handled.

task-6214240
runbot-error-242920

X-original-commit: a200d0d361f2d2f64268e279889527f474ac9ab3
