+++ addons/pos_stock/models/pos_session.py
@@ -1,7 +1,7 @@
 from collections import defaultdict
 
 from odoo import api, fields, models
-from odoo.tools import float_compare, float_is_zero, split_every
+from odoo.tools import split_every
 from odoo.tools.constants import PREFETCH_MAX
 
 
@@ -79,33 +79,13 @@ class PosSession(models.Model):
         data = self._create_stock_valuation_lines(data)
         return data
 
-    def _get_rounding_difference_vals(self, amount, amount_converted):
-        if self.config_id.cash_rounding:
-            partial_args = {
-                'name': 'Rounding line',
-                'move_id': self.move_id.id,
-            }
-            if float_compare(0.0, amount, precision_rounding=self.currency_id.rounding) > 0:    # loss
-                partial_args['account_id'] = self.config_id.rounding_method.loss_account_id.id
-                return self._debit_amounts(partial_args, -amount, -amount_converted)
-
-            if float_compare(0.0, amount, precision_rounding=self.currency_id.rounding) < 0:   # profit
-                partial_args['account_id'] = self.config_id.rounding_method.profit_account_id.id
-                return self._credit_amounts(partial_args, amount, amount_converted)
-
     def _create_non_reconciliable_move_lines(self, data):
         data = super()._create_non_reconciliable_move_lines(data)
         stock_expense = data.get('stock_expense')
-        rounding_difference = data.get('rounding_difference')
         MoveLine = data.get('MoveLine')
-        rounding_vals = []
-
-        if not float_is_zero(rounding_difference['amount'], precision_rounding=self.currency_id.rounding) or not float_is_zero(rounding_difference['amount_converted'], precision_rounding=self.currency_id.rounding):
-            rounding_vals = [self._get_rounding_difference_vals(rounding_difference['amount'], rounding_difference['amount_converted'])]
 
         MoveLine.create(
             [self._get_stock_expense_vals(key, amounts['amount'], amounts['amount_converted']) for key, amounts in stock_expense.items()]
-            + rounding_vals
         )
 
         return data
